mini refactoring fixed texture resolution
