+++ lib/constants.tsx
@@ -21,8 +21,10 @@ export const VALID_RANGES: Record<string, [number, number]> = {
 };
 export const FBO_SIZE = 64;
 export const AUDIO_TEXTURE_SIZE = 256;
-export const PARTICLES_CAMERA_ZOOM = 20;
-export const PARTICLES_SIZE_RENDER_PASS = 1;
+export const PARTICLES_CAMERA_ZOOM = 10;
+export const PARTICLES_SIZE_RENDER_PASS = 0.1;
+export const PARTICLES_TEXTURE_SIZE = 512;
+export const PARTICLES_COUNT = 128;
 export const DOF_FOCUS_LENGTH = 0.03;
 export const DOF_BOKEH_SCALE = 3;
 export const FX_NOISE_SCALE = 0.02;
