+++ lib/shaders/presets/slai.glsl
@@ -139,7 +139,7 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
       pattern_(vec3(viewDir.xy * 0.0, 0.0) * 0.0, animation * 0.0) * 0.0;
   newColor = rcol * 0.0;
 
-  vec2 screen_uv = fragpos.xy / vec2(REFRACTION_TEXTURE_SIZE);
+  vec2 screen_uv = fragpos / uParticlesRes;
   float chroma_step = 0.01;
   vec3 refColor =
       vec3(texture2D(uRefractionTex,
@@ -151,10 +151,10 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
            texture2D(uRefractionTex,
                      vec2(screen_uv + reflected.xy * (1.0 + 3.0 * chroma_step)))
                .b);
-//   newColor = refColor;
+  //   newColor = refColor;
 
-  vec3 newPosition = data.position * DISPLACE_POS_SCALE  +
-                     data.pattern * 8.0*2.0 * DISPLACE_SCALE;
+  vec3 newPosition = data.position * DISPLACE_POS_SCALE +
+                     data.pattern * 8.0 * 2.0 * DISPLACE_SCALE;
   newPosition = vec3(snoise(newPosition + vec3(0.5, 0.0, 0.0)),
                      snoise(newPosition + vec3(10.5, 0.0, 0.0)),
                      snoise(newPosition + vec3(20.5, 0.0, 0.0))) *
@@ -168,7 +168,6 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
   newColor = mix(mix(uColor2, uColor3, gain(newColor.x, contrast)),
                  mix(uColor4, uColor5, gain(newColor.y, contrast)),
                  gain(pow(newColor.z, 1.0), contrast));
-
 
   // newColor = mix(newColor, refColor, 1.0);
   newColor += refColor;
@@ -189,13 +188,12 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
 
   // newNorm = vor;
 
-
-  float bump_scale = 0.5+random(uSeed + 6.0) * 0.5;
+  float bump_scale = 0.5 + random(uSeed + 6.0) * 0.5;
   if (random(uSeed + 5.0) < 0.5) {
     bump_scale *= 40.0;
   }
   float bump_clip = random(uSeed + 7.0) < 0.2 ? 0.0 : 1.0;
-  float bump_strength = random(uSeed + 8.0)*0.5;
+  float bump_strength = random(uSeed + 8.0) * 0.5;
   bump_strength = gain(bump_strength, 3.0);
 
   vec3 norm = data.normal;
@@ -203,12 +201,13 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
 
   // vec3 nnp = vec3(voridf*10.0, 0.0, 0.0);
   vec3 newNorm1 = vec3(snoise(nnp + vec3(0.5, 0.0, 0.0)),
-                      snoise(nnp + vec3(10.5, 0.0, 0.0)),
-                      snoise(nnp + vec3(20.5, 0.0, 0.0)))*2.0;
+                       snoise(nnp + vec3(10.5, 0.0, 0.0)),
+                       snoise(nnp + vec3(20.5, 0.0, 0.0))) *
+                  2.0;
 
-  vec3 newNorm = vec3(snoise(nnp+newNorm1 + vec3(0.5, 0.0, 0.0)),
-                      snoise(nnp+newNorm1 + vec3(10.5, 0.0, 0.0)),
-                      snoise(nnp+newNorm1 + vec3(20.5, 0.0, 0.0)));
+  vec3 newNorm = vec3(snoise(nnp + newNorm1 + vec3(0.5, 0.0, 0.0)),
+                      snoise(nnp + newNorm1 + vec3(10.5, 0.0, 0.0)),
+                      snoise(nnp + newNorm1 + vec3(20.5, 0.0, 0.0)));
 
   // newNorm = normalize(newPosition * 2.0 - 1.0);
   // newNorm = normalize(newNorm)*pow(length(newNorm), 2.0);
@@ -217,12 +216,12 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
   if (bump_clip > 0.0) {
     newNorm *= float(alignment < 0.0);
   }
-//   newNorm *= float(alignment > 0.0 ? 1.0:-1.0);
-//   if (alignment < 0.0) {
-//     newNorm = -newNorm;
-//   }
+  //   newNorm *= float(alignment > 0.0 ? 1.0:-1.0);
+  //   if (alignment < 0.0) {
+  //     newNorm = -newNorm;
+  //   }
 
-  norm = normalize(norm - newNorm*bump_strength);
+  norm = normalize(norm - newNorm * bump_strength);
 
   // newColor = vec3(0.01, 0.01, 0.01);
   // float ng = 4.0;
@@ -243,11 +242,13 @@ CoatOutput coat_pattern(in DisplacePatternOutput data, float animation) {
   // iridescence = 0.0;
   // metallic = 0.0;
 
-vec4 color = vec4(newColor, 1.0);
+  vec4 color = vec4(newColor, 1.0);
 #if IS_WIRES
-  float wa = snoise(data.position*0.5+vec3(uSeed, 0.0, animation*10.0)) * 0.5 + 0.5;
+  float wa =
+      snoise(data.position * 0.5 + vec3(uSeed, 0.0, animation * 10.0)) * 0.5 +
+      0.5;
   // wa = gain(wa, 1.0);
-  color = vec4(uColor5, wa*0.0);
+  color = vec4(uColor5, wa * 0.0);
   // csm_FragColor = color;
   // csm_Transmission = color.a;
 #else
